add docs

+++ seeds/populateTasks.js
@@ -1,20 +1,33 @@
 const connectDB = require('../config/db')
-const taskData = require('./tasks.json')
 const fs = require('fs')
 const Task = require('../schema/taskSchema')
+require('dotenv').config()
+
+/*
+  *
+  * Usage: From the root project directory run $ node ./seeds/populateTasks.js
+  * Overview: Clears current tasks in DB and repopulates with data from tasks.json (100 tasks)
+  * Note: Will not run in production
+*/
 
 function populateTasks() {
 
   connectDB()
 
-  deleteTasks().then((res) => {
-    console.log("deleted tasks successfully")
-    insertTasks()
-      .then((res) => console.log("inserted tasks successfully"))
-      .catch((err) => {
-        console.log("error inserting tasks: ", err)
-      }).catch((err) => console.log("error deleted tasks: ", err))
-  })
+  if (process.env.NODE_ENV !== 'production') {
+    deleteTasks().then((res) => {
+      console.log("deleted tasks successfully")
+      insertTasks()
+        .then((res) => console.log("inserted tasks successfully"))
+        .catch((err) => {
+          console.log("error inserting tasks: ", err)
+        }).catch((err) => console.log("error deleted tasks: ", err))
+    })
+
+  } else {
+    console.log("You cannot do this in production environment")
+    process.exit(1)
+  }
 }
 
 async function insertTasks() {
@@ -29,7 +42,9 @@ async function insertTasks() {
 
 async function deleteTasks() {
 
+  // Wipe current tasks
   const deletedTasks = await Task.deleteMany({})
+
   return deletedTasks
 }
